app/requirements/base.txt
asgiref==3.3.1
Django==3.1.3
django-filter==2.4.0
django-ordered-model==3.4.1
djangorestframework==3.12.2
greenlet==0.4.17
Markdown==3.3.3
msgpack==1.0.0
Pillow==8.0.1
pynvim==0.4.2
pytz==2020.4
sqlparse==0.4.1
